refactor: split pkg/runtime/docker/docker.go - lifecycle.go + exec.go (issue 3, part B)
